+++ backend/src/main/java/com/ktnu/AiLectureSummary/controller/AuthController.java
@@ -2,9 +2,9 @@ package com.ktnu.AiLectureSummary.controller;
 
 import com.ktnu.AiLectureSummary.config.CookieProperties;
 import com.ktnu.AiLectureSummary.dto.ApiResponse;
-import com.ktnu.AiLectureSummary.dto.member.MemberLoginRequest;
-import com.ktnu.AiLectureSummary.dto.member.MemberLoginResponse;
-import com.ktnu.AiLectureSummary.dto.member.MemberRegisterRequest;
+import com.ktnu.AiLectureSummary.dto.member.request.MemberLoginRequest;
+import com.ktnu.AiLectureSummary.dto.member.response.MemberLoginResponse;
+import com.ktnu.AiLectureSummary.dto.member.request.MemberRegisterRequest;
 import com.ktnu.AiLectureSummary.service.MemberAuthService;
 import com.ktnu.AiLectureSummary.util.CookieResponseUtil;
 import io.swagger.v3.oas.annotations.Operation;
