+++ backend/src/main/java/com/ktnu/AiLectureSummary/controller/ProfileController.java
@@ -3,9 +3,9 @@ package com.ktnu.AiLectureSummary.controller;
 import com.ktnu.AiLectureSummary.application.MemberDeleteApplicationService;
 import com.ktnu.AiLectureSummary.config.CookieProperties;
 import com.ktnu.AiLectureSummary.dto.ApiResponse;
-import com.ktnu.AiLectureSummary.dto.member.MemberEditRequest;
-import com.ktnu.AiLectureSummary.dto.member.MemberEditResponse;
-import com.ktnu.AiLectureSummary.dto.member.MemberProfileResponse;
+import com.ktnu.AiLectureSummary.dto.member.request.MemberEditRequest;
+import com.ktnu.AiLectureSummary.dto.member.response.MemberEditResponse;
+import com.ktnu.AiLectureSummary.dto.member.response.MemberProfileResponse;
 import com.ktnu.AiLectureSummary.security.CustomUserDetails;
 import com.ktnu.AiLectureSummary.service.MemberProfileService;
 import com.ktnu.AiLectureSummary.util.CookieResponseUtil;
