+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/lecture/request/YoutubeLectureRequest.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.lecture;
+package com.ktnu.AiLectureSummary.dto.lecture.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
