+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/member/request/MemberRegisterRequest.java
@@ -1,14 +1,18 @@
-package com.ktnu.AiLectureSummary.dto.member;
+package com.ktnu.AiLectureSummary.dto.member.request;
 
+import com.ktnu.AiLectureSummary.constant.ValidationRegex;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
-import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+
+/**
+ * 회원가입 요청 DTO
+ */
 @Getter
 @Setter
 @AllArgsConstructor
@@ -19,12 +23,14 @@ public class MemberRegisterRequest {
     @Email(message = "유효한 이메일 주소여야 합니다.")
     private String email;
 
+    @Schema(description = "사용자 이름", example = "testUser")
+    @Pattern(regexp = ValidationRegex.USERNAME, message = "이름에는 연속된 공백을 포함할 수 없으며, 최대 30자까지 가능합니다.")
+    private String username;
+
     @Schema(description = "사용자 비밀번호", example = "testpassword123!")
     @NotBlank(message = "비밀번호는 비어 있을 수 없습니다.")
-    @Pattern(regexp = "^(?=\\S+$).{8,20}$", message = "비밀번호는 공백 없이 8자 이상 20자 이하여야 합니다.")
+    @Pattern(regexp = ValidationRegex.PASSWORD, message = "비밀번호는 공백 없이 8자 이상 20자 이하여야 합니다.")
     private String password;
 
-    @Schema(description = "사용자 이름", example = "testUser")
-    @Pattern(regexp = "^(?!.*\\s{2,}).{1,30}$", message = "이름에는 연속된 공백을 포함할 수 없으며, 최대 30자까지 가능합니다.")
-    private String username;
+
 }
