+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/member/request/MemberVerifyRequest.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.member;
+package com.ktnu.AiLectureSummary.dto.member.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.Email;
@@ -7,6 +7,9 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+/**
+ * 비밀번호 재설정 전 본인 인증 요청 DTO
+ */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
