+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/member/response/MemberEditResponse.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.member;
+package com.ktnu.AiLectureSummary.dto.member.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
