+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/member/response/MemberProfileResponse.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.member;
+package com.ktnu.AiLectureSummary.dto.member.response;
 
 
 import com.ktnu.AiLectureSummary.security.CustomUserDetails;
