+++ backend/src/main/java/com/ktnu/AiLectureSummary/dto/memberLecture/response/MemberLectureListItemResponse.java
@@ -1,4 +1,4 @@
-package com.ktnu.AiLectureSummary.dto.memberLecture;
+package com.ktnu.AiLectureSummary.dto.memberLecture.response;
 
 import com.ktnu.AiLectureSummary.domain.MemberLecture;
 import com.ktnu.AiLectureSummary.util.ThumbnailUtil;
