+++ backend/src/main/java/com/ktnu/AiLectureSummary/service/MemberAuthService.java
@@ -3,7 +3,9 @@ package com.ktnu.AiLectureSummary.service;
 import com.ktnu.AiLectureSummary.config.JwtProperties;
 import com.ktnu.AiLectureSummary.domain.Member;
 import com.ktnu.AiLectureSummary.domain.Role;
-import com.ktnu.AiLectureSummary.dto.member.*;
+import com.ktnu.AiLectureSummary.dto.member.request.MemberLoginRequest;
+import com.ktnu.AiLectureSummary.dto.member.request.MemberRegisterRequest;
+import com.ktnu.AiLectureSummary.dto.member.response.MemberLoginResponse;
 import com.ktnu.AiLectureSummary.exception.*;
 import com.ktnu.AiLectureSummary.repository.MemberRepository;
 import com.ktnu.AiLectureSummary.security.JwtProvider;
