+++ backend/src/main/java/com/ktnu/AiLectureSummary/service/MemberLectureService.java
@@ -1,8 +1,7 @@
 package com.ktnu.AiLectureSummary.service;
 
 import com.ktnu.AiLectureSummary.domain.Member;
-import com.ktnu.AiLectureSummary.exception.PdfGenerateFailException;
-import com.ktnu.AiLectureSummary.dto.memberLecture.MemberLectureListResponse;
+import com.ktnu.AiLectureSummary.dto.memberLecture.response.MemberLectureListResponse;
 import com.ktnu.AiLectureSummary.exception.MemberNotFoundException;
 import com.ktnu.AiLectureSummary.repository.LectureRepository;
 import com.ktnu.AiLectureSummary.repository.MemberRepository;
@@ -10,19 +9,14 @@ import com.ktnu.AiLectureSummary.repository.MemberRepository;
 
 import com.ktnu.AiLectureSummary.domain.Lecture;
 import com.ktnu.AiLectureSummary.domain.MemberLecture;
-import com.ktnu.AiLectureSummary.dto.memberLecture.LectureDetailResponse;
+import com.ktnu.AiLectureSummary.dto.memberLecture.response.LectureDetailResponse;
 import com.ktnu.AiLectureSummary.exception.LectureNotFoundException;
 import com.ktnu.AiLectureSummary.repository.MemberLectureRepository;
 import com.ktnu.AiLectureSummary.security.CustomUserDetails;
-import com.lowagie.text.Document;
-import com.lowagie.text.DocumentException;
-import com.lowagie.text.Paragraph;
-import com.lowagie.text.pdf.PdfWriter;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
-import java.io.ByteArrayOutputStream;
 import java.util.List;
 
 @Service
